safer_imsave() to util
